% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_themes.R
\name{theme_super_minimal}
\alias{theme_super_minimal}
\title{Blog}
\usage{
theme_super_minimal(base_family = "Roboto", fill = "#fffff8", ...)
}
\arguments{
\item{base_family}{font type, default to Roboto}

\item{fill}{default to #fffff8, which is the color in tufte handouts}

\item{...}{}
}
\value{
a custom ggplot2 theme
}
\description{
Blog
}
